+++ FrontEnd/src/utils/numberUtils.js
@@ -2,8 +2,6 @@
 
 
 /**
- * FormatPrice
- *
  * Converts a price value into a formatted string for display.
  *
  * @param {number} price - The price value to format.
@@ -18,4 +16,16 @@ export function formatPrice(price) {
       const roundedPrice = parseFloat(price.toPrecision(4))
       return `$${roundedPrice}`;
   }
+}
+
+/**
+ * Converts a time format into a convient time string for display
+ *
+ * @param {number} price - The time value to format.
+ * @returns {string} The formatted time string.
+ */
+export function formatTime(number) {
+  const scale = 100
+  const duration = Math.round(number * scale) / scale
+  return `${duration}s`;
 }