+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New users are sent email verification emails
   - New user *can* receive the free dollar promotion after verification
 - Newly created messages will be automatically selected by default
+- Workflow duration in seconds
 
 ### Changed